.history/requirements_20251217163403.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
pillow
gunicorn
